hw9/requirements.txt
google-cloud-storage==2.18.2
google-cloud-logging==3.11.3
requests==2.32.3
